+++ src/resources/views/components/universal-priority.blade.php
@@ -1,0 +1,3 @@
+<universal-priority url="{{ route("ajax.priority.update", ["table" => $table, "field" => $field]) }}"
+                    :elements="{{ json_encode($elements) }}">
+</universal-priority>