+++ src/routes/priority.php
@@ -1,0 +1,16 @@
+<?php
+
+use Illuminate\Support\Facades\Route;
+use MBober35\Helpers\Http\Controllers\PriorityController;
+
+Route::group([
+    "prefix" => "ajax/vue/priority",
+    "middleware" => ["web", "management"],
+    "as" => "ajax.priority."
+], function () {
+    Route::group([
+        "prefix" => "/{table}/{field}",
+    ], function () {
+        Route::put("/", [PriorityController::class, "update"])->name("update");
+    });
+});