+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # ChangeLog
 
+## [1.4.0]
+- Добавлен компонент для изменения приоритета.
+### Обновление:
+- php artisan vendor:publish --provider="MBober35\Helpers\ServiceProvider" --tag=public
+- php artisan helpers:init (Добавить universal priority)
+
 ## [1.3.0]
 - Расширен компонент меню, что бы вывести меню для темы AdminKit